html/browsers/browsing-the-web/navigating-across-documents/009-original.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#concept-form-submit">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-get-action">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#plan-to-navigate">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#planned-navigation">
<title>Link with onclick form submit to javascript: url with document.write and href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<form target="test" action="javascript:(() => {document.write(`<script>parent.postMessage('write', '*')</script>`); return `<script>parent.postMessage('click', '*')</script>`})()"></form>
<a target="test" onclick="document.forms[0].submit()" href="href.html">Test</a>
<script>
var test = async_test();
var messages = [];

test.step(function() {
  document.querySelector("a").click();
});

window.onmessage = test.step_func_done(function(e) {
  assert_equals(e.data, "href");
});
</script>

<!--
1. This test creates an iframe, a form with
    `action="javascript:document.write('<script>parent.postMessage(&quot;write&quot;, &quot;*&quot;)</script>');"`,
    and an anchor with both an `onclick` and `href`.
2. The test "clicks" the anchor, triggering the `onclick` handler operation,
    which will attempt to submit the form. The form's `action` is a
    "javascript: url", which has special handling in step 12 (https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents)
3. Navigating the iframe to the anchor's `href` will abort the `action`.
4. If the semantics are implemented correctly, a message "href"
    from "href.html" will be received.

-->
